fix error in expected iterations

+++ src/simulation/interceptors/particle_update_counter/ParticleUpdateCounterInterceptor.cpp
@@ -11,9 +11,9 @@ void ParticleUpdateCounterInterceptor::onSimulationStart() {
     t_start = std::chrono::high_resolution_clock::now();
 }
 
-void ParticleUpdateCounterInterceptor::operator()(int iteration) { particle_updates += simulation.particle_container->size(); }
+void ParticleUpdateCounterInterceptor::operator()(size_t iteration) { particle_updates += simulation.particle_container->size(); }
 
-void ParticleUpdateCounterInterceptor::onSimulationEnd(int iteration) {
+void ParticleUpdateCounterInterceptor::onSimulationEnd(size_t iteration) {
     t_end = std::chrono::high_resolution_clock::now();
     t_diff = std::chrono::duration_cast<std::chrono::milliseconds>(t_end - t_start).count();
 
@@ -22,11 +22,9 @@ void ParticleUpdateCounterInterceptor::onSimulationEnd(int iteration) {
 
 double ParticleUpdateCounterInterceptor::getParticleUpdatesPerSecond() const { return particle_updates_per_second; }
 
-
 std::chrono::milliseconds::rep ParticleUpdateCounterInterceptor::getSimulationDurationMS() const {
     return std::chrono::duration_cast<std::chrono::milliseconds>(t_end - t_start).count();
 }
-
 
 ParticleUpdateCounterInterceptor::operator std::string() const {
     return "ParticleUpdateInterceptor: " + std::to_string(particle_updates_per_second) + " particle updates per second";
